requirements.txt
PyPDF2==3.0.1
pdfminer.six==20221105
boto3==1.34.69
markdown==3.5.2
pyyaml==6.0.1
